First version of RandomDungeonGenerator is added.
This is a random dungeon generation script that exposes a method as an API which creates a random room matrix.
Known issues:
When the total room count is too close to the whole size of the matrix (width * height) the algorithm can get stuck in the room generation loop (probable cause: algorithm cannot find the rooms that lead to the few rooms that it still needs to set).