Inventory updated.
Since the inventory script was initially designed to work with object pools and that wasted memory since in most scenarios an inventory doesn't need to have a pool.
Name of the Inventory script and class is now changed to PooledInventory.
Unit tests are added for PooledInventory  that tests every public method (currently 36 tests).
An inventory script that has the exact same interface but works without an object pool is going to be implemented and uploaded.