Non-Pooled inventory added.
This inventory doesn't work with object pools unlike the PooledInventory.
PooledInventory and Inventory has the exact same interface, when the user want to change from the other, all they need to do is change the type name.
Unit tests for inventory are also added. These are the same tests with PooledInventory since they share the exact same interface, only the type names are different.
TODO: When unity supports C# 9.0, inventory is going to return null from the indexer when there is no item in the slot instead of throwing an exception.