Added FetchAttribute to eliminate the serialization in inspector and GetComponent calls.
The way this attribute works is that it needs to be put on a field
inside a MonoBehaviour inherited class on a gameobject. Then, whenever
the fields need to be initialized (on Awake() or on Start() for
example),
RunFetch(this);
should be run which will initialize all the fields using the components
on the gameobject.